+++ src/kernel_evm/kernel/src/inbox.rs
@@ -6,6 +6,7 @@
 // SPDX-License-Identifier: MIT
 
 use crate::error::UpgradeProcessError::InvalidUpgradeNonce;
+use crate::outbox_counter::reset_outbox_counter;
 use crate::parsing::{Input, InputResult, MAX_SIZE_PER_CHUNK, UPGRADE_NONCE_SIZE};
 use crate::simulation;
 use crate::storage::{
@@ -320,6 +321,7 @@ pub fn read_inbox<Host: Runtime>(
                 });
             }
             InputResult::Input(Input::Info(info)) => {
+                reset_outbox_counter(host)?;
                 store_last_info_per_level_timestamp(host, info.predecessor_timestamp)?;
             }
             InputResult::Input(Input::Deposit(deposit)) => {